Fix #183 - E022 - Improve text for stop_time_update validation
* Improve text for E022 so "equals" and "less than" are split out as separate cases
* Add times in human-readable format ("09:23:98") to the occurrence summary
* New rule E025 for comparing arrival and departure time within the same stop_time_update
* Move unit test previously under E022 to E025 method and fix unit tests for E022 so they pass without this code
* Add agency.txt agency_time_zone to GTFS Metadata - needed to calculate correct times in clock format like 09:23:09
* Fix/improve comments

+++ src/main/java/edu/usf/cutr/gtfsrtvalidator/util/TimestampUtils.java
@@ -75,7 +75,7 @@ public class TimestampUtils {
      * @param timeZone the timezone used to generate the clock time, or null if the current time zone should be used.  Please refer to http://en.wikipedia.org/wiki/List_of_tz_zones for a list of valid values.
      * @return A converted version of time in 24hr clock time like "06:00:00"
      */
-    public static String posixToClock(int posixTime, TimeZone timeZone) {
+    public static String posixToClock(long posixTime, TimeZone timeZone) {
         if (timeZone != null) {
             mTimeFormat.setTimeZone(timeZone);
         }
